Add security package and move CustomUserDetails and SecurityConfig there
