+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/config/security/SecurityConfig.java
@@ -1,6 +1,5 @@
-package com.avdo.spring.app.config;
+package com.avdo.spring.app.config.security;
 
-import com.avdo.spring.app.entity.CustomUserDetails;
 import com.avdo.spring.app.filter.JwtAuthenticationFilter;
 import com.avdo.spring.app.repository.UserRepository;
 import com.avdo.spring.app.service.domain.model.User;
@@ -16,14 +15,11 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-import java.util.Collections;
 
 
 @Configuration
